Try a new scheme for updating user's keys to avoid locking them out when over quota.
The old method was to write the key data directly to the authorized_keys files, then ensure that it was owned by the right user. If it already existed and was owned by that user, when the user was over quota the attempt to overwrite it would apparently successfully truncate it, but be blocked from writing any new data, leaving the user locked out. he new scheme is to always write to a different file, which should not generally exist, which will then be created owned by root. This file is then chowned to the correct user, and moved into place.